+++ src/shared/infra/typeorm/migrations/1666223297687-CreateBullFeatures.ts
@@ -20,7 +20,7 @@ export class CreateBullFeatures1666223297687 implements MigrationInterface {
                         name: "profile_item_id",
                     },
                     {
-                        type: "numeric",
+                        type: "float",
                         name: "value",
                     },
                 ],
